requirement.txt
torch
transformers
Pillow
sentence-transformers
faiss-cpu
